+++ app/ante/cosmos/authz.go
@@ -13,7 +13,7 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Evmos packages. If not, see https://github.com/evmos/evmos/blob/main/LICENSE
-package evm
+package cosmos
 
 import (
 	"fmt"
