+++ app/ante/cosmos/authz_test.go
@@ -1,4 +1,4 @@
-package evm_test
+package cosmos_test
 
 import (
 	"fmt"
@@ -13,7 +13,7 @@ import (
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
-	evmante "github.com/evmos/evmos/v11/app/ante/evm"
+	cosmosante "github.com/evmos/evmos/v11/app/ante/cosmos"
 	evmtypes "github.com/evmos/evmos/v11/x/evm/types"
 )
 
@@ -30,7 +30,7 @@ func TestAuthzLimiterDecorator(t *testing.T) {
 	stakingAuthUndelegate, err := stakingtypes.NewStakeAuthorization([]sdk.ValAddress{validator}, nil, stakingtypes.AuthorizationType_AUTHORIZATION_TYPE_UNDELEGATE, nil)
 	require.NoError(t, err)
 
-	decorator := evmante.NewAuthzLimiterDecorator(
+	decorator := cosmosante.NewAuthzLimiterDecorator(
 		sdk.MsgTypeURL(&evmtypes.MsgEthereumTx{}),
 		sdk.MsgTypeURL(&stakingtypes.MsgUndelegate{}),
 	)
