+++ app/ante/cosmos/utils_test.go
@@ -10,17 +10,28 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/signer/core/apitypes"
 
 	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/tx"
+	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/simapp"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/tx/signing"
+	"github.com/cosmos/cosmos-sdk/x/auth/legacy/legacytx"
+	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	"github.com/cosmos/cosmos-sdk/x/authz"
 
 	"github.com/evmos/evmos/v11/app"
 	ante "github.com/evmos/evmos/v11/app/ante"
+	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
 	"github.com/evmos/evmos/v11/encoding"
 	"github.com/evmos/evmos/v11/ethereum/eip712"
+	"github.com/evmos/evmos/v11/tests"
+	"github.com/evmos/evmos/v11/types"
 	"github.com/evmos/evmos/v11/utils"
 	"github.com/evmos/evmos/v11/x/evm/statedb"
 	evmtypes "github.com/evmos/evmos/v11/x/evm/types"
@@ -138,3 +149,170 @@ type invalidTx struct{}
 func (invalidTx) GetMsgs() []sdk.Msg { return []sdk.Msg{nil} }
 
 func (invalidTx) ValidateBasic() error { return nil }
+
+func newMsgExec(grantee sdk.AccAddress, msgs []sdk.Msg) *authz.MsgExec {
+	msg := authz.NewMsgExec(grantee, msgs)
+	return &msg
+}
+
+func newMsgGrant(granter sdk.AccAddress, grantee sdk.AccAddress, a authz.Authorization, expiration *time.Time) *authz.MsgGrant {
+	msg, err := authz.NewMsgGrant(granter, grantee, a, expiration)
+	if err != nil {
+		panic(err)
+	}
+	return msg
+}
+
+func createNestedMsgExec(a sdk.AccAddress, nestedLvl int, lastLvlMsgs []sdk.Msg) *authz.MsgExec {
+	msgs := make([]*authz.MsgExec, nestedLvl)
+	for i := range msgs {
+		if i == 0 {
+			msgs[i] = newMsgExec(a, lastLvlMsgs)
+			continue
+		}
+		msgs[i] = newMsgExec(a, []sdk.Msg{msgs[i-1]})
+	}
+	return msgs[nestedLvl-1]
+}
+
+func generatePrivKeyAddressPairs(accCount int) ([]*ethsecp256k1.PrivKey, []sdk.AccAddress, error) {
+	var (
+		err           error
+		testPrivKeys  = make([]*ethsecp256k1.PrivKey, accCount)
+		testAddresses = make([]sdk.AccAddress, accCount)
+	)
+
+	for i := range testPrivKeys {
+		testPrivKeys[i], err = ethsecp256k1.GenerateKey()
+		if err != nil {
+			return nil, nil, err
+		}
+		testAddresses[i] = testPrivKeys[i].PubKey().Address().Bytes()
+	}
+	return testPrivKeys, testAddresses, nil
+}
+
+func createTx(priv *ethsecp256k1.PrivKey, msgs ...sdk.Msg) (sdk.Tx, error) {
+	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
+	txBuilder := encodingConfig.TxConfig.NewTxBuilder()
+
+	txBuilder.SetGasLimit(1000000)
+	if err := txBuilder.SetMsgs(msgs...); err != nil {
+		return nil, err
+	}
+
+	// First round: we gather all the signer infos. We use the "set empty
+	// signature" hack to do that.
+	sigV2 := signing.SignatureV2{
+		PubKey: priv.PubKey(),
+		Data: &signing.SingleSignatureData{
+			SignMode:  encodingConfig.TxConfig.SignModeHandler().DefaultMode(),
+			Signature: nil,
+		},
+		Sequence: 0,
+	}
+
+	sigsV2 := []signing.SignatureV2{sigV2}
+
+	if err := txBuilder.SetSignatures(sigsV2...); err != nil {
+		return nil, err
+	}
+
+	signerData := authsigning.SignerData{
+		ChainID:       "evmos_9000-1",
+		AccountNumber: 0,
+		Sequence:      0,
+	}
+	sigV2, err := tx.SignWithPrivKey(
+		encodingConfig.TxConfig.SignModeHandler().DefaultMode(), signerData,
+		txBuilder, priv, encodingConfig.TxConfig,
+		0,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	sigsV2 = []signing.SignatureV2{sigV2}
+	err = txBuilder.SetSignatures(sigsV2...)
+	if err != nil {
+		return nil, err
+	}
+
+	return txBuilder.GetTx(), nil
+}
+
+func createEIP712CosmosTx(
+	from sdk.AccAddress, priv cryptotypes.PrivKey, msgs []sdk.Msg,
+) (sdk.Tx, error) {
+	var err error
+
+	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
+	txBuilder := encodingConfig.TxConfig.NewTxBuilder()
+
+	// GenerateTypedData TypedData
+	registry := codectypes.NewInterfaceRegistry()
+	types.RegisterInterfaces(registry)
+	ethermintCodec := codec.NewProtoCodec(registry)
+	cryptocodec.RegisterInterfaces(registry)
+
+	coinAmount := sdk.NewCoin(evmtypes.DefaultEVMDenom, sdk.NewInt(20))
+	amount := sdk.NewCoins(coinAmount)
+	gas := uint64(200000)
+
+	fee := legacytx.NewStdFee(gas, amount)
+	data := legacytx.StdSignBytes("evmos_9000-1", 0, 0, 0, fee, msgs, "", nil)
+	typedData, err := eip712.WrapTxToTypedData(ethermintCodec, 9000, msgs[0], data, &eip712.FeeDelegationOptions{
+		FeePayer: from,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	sigHash, _, err := apitypes.TypedDataAndHash(typedData)
+	if err != nil {
+		return nil, err
+	}
+
+	// Sign typedData
+	keyringSigner := tests.NewSigner(priv)
+	signature, pubKey, err := keyringSigner.SignByAddress(from, sigHash)
+	if err != nil {
+		return nil, err
+	}
+	signature[crypto.RecoveryIDOffset] += 27 // Transform V from 0/1 to 27/28 according to the yellow paper
+
+	// Add ExtensionOptionsWeb3Tx extension
+	var option *codectypes.Any
+	option, err = codectypes.NewAnyWithValue(&types.ExtensionOptionsWeb3Tx{
+		FeePayer:         from.String(),
+		TypedDataChainID: 9000,
+		FeePayerSig:      signature,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	builder, _ := txBuilder.(authtx.ExtensionOptionsTxBuilder)
+
+	builder.SetExtensionOptions(option)
+	builder.SetFeeAmount(amount)
+	builder.SetGasLimit(gas)
+
+	sigsV2 := signing.SignatureV2{
+		PubKey: pubKey,
+		Data: &signing.SingleSignatureData{
+			SignMode: signing.SignMode_SIGN_MODE_LEGACY_AMINO_JSON,
+		},
+		Sequence: 0,
+	}
+
+	if err = builder.SetSignatures(sigsV2); err != nil {
+		return nil, err
+	}
+
+	if err = builder.SetMsgs(msgs...); err != nil {
+		return nil, err
+	}
+
+	return builder.GetTx(), err
+}
